Snakes and Ladders V.2
Changed some of the things in the game rules. The board actually prints now but I have to figure out a way to incorporate snakes and ladders on the board also players.  I still have to make the dice accessible to the user and create the option of having to go against a player or "AI".