Working on the player movements and fixing errors.
I've created a method for manual location of the snakes and ladders instead of keeping it in the same method of where the game board prints due to it having to reset if the player lands on the location of a snake or ladder. There are some tiny changes in methods (comments, fixes, etc.) I am currently working on the player movements once, I am finished it will be easy to create the second player and AI.  My worries right now are the movements fixing the movements of the player. Once I am done this I will continue fixing the problem of players landing on the same location on the board.